Wenjing

+++ project/src/main/java/com/course/demo/Service/TeacherService.java
@@ -34,7 +34,11 @@ public class TeacherService {
     //2.添加老师
     //添加一个老师
     public boolean addTeacher(TeacherEntity teacherEntity){
-
+        //如果老师已存在，则添加失败
+        if(teacherDao.findById(teacherEntity.getId())!=null){
+            System.out.println("添加老师出错，该老师已存在");
+            return false;
+        }
         try
         {
             teacherDao.save(teacherEntity);
@@ -47,6 +51,10 @@ public class TeacherService {
     }
     //3.删除老师
     public boolean removeTeacher(int id){
+        if(teacherDao.findById(id)==null){
+            System.out.println("删除错误，该老师不存在");
+            return false;
+        }
         try {
             teacherDao.deleteById(id);
         }
@@ -57,4 +65,79 @@ public class TeacherService {
         return  true;
     }
     //4.修改老师信息
+    //修改老师姓名,找到id对应的老师，按name更新
+    public boolean updateTeacherNameById(int id,String name){
+        //看老师是否存在，不存在报错
+        if(teacherDao.findById(id)==null){
+            System.out.println("修改错误，该老师不存在");
+            return false;
+        }
+        //通过id找到，修改对应信息
+        try{
+            TeacherEntity teacherEntity=teacherDao.findById(id);
+            teacherEntity.setName(name);
+            teacherDao.save(teacherEntity);
+        }
+        catch (Exception e){
+            System.out.println("修改老师姓名出错："+e.getCause()+" "+e.getMessage());
+            return  false;
+        }
+        return true;
+    }
+    //修改老师头像，找到id对应的老师，按teacher_url更新
+    public boolean updateTeacherUrlById(int id,String teacher_url){
+        if(teacherDao.findById(id)==null){
+            System.out.println("修改错误，该老师不存在");
+            return false;
+        }
+        //通过id找到，修改对应信息
+        try {
+            TeacherEntity teacherEntity=teacherDao.findById(id);
+            teacherEntity.setTeacherUrl(teacher_url);
+            teacherDao.save(teacherEntity);
+        }
+        catch (Exception e){
+            System.out.println("修改老师头像错误："+e.getCause()+" "+e.getMessage());
+            return false;
+        }
+        return true;
+    }
+    //修改老师介绍，找到id对应的老师，按intro更新
+    public boolean updateTeacherIntroById(int id,String intro){
+        //看老师是否存在，不存在报错
+        if(teacherDao.findById(id)==null){
+            System.out.println("修改错误，该老师不存在");
+            return false;
+        }
+        //通过id找到，修改对应信息
+        try {
+            TeacherEntity teacherEntity=teacherDao.findById(id);
+            teacherEntity.setIntro(intro);
+            teacherDao.save(teacherEntity);
+        }
+        catch (Exception e){
+            System.out.println("修改老师介绍出错"+e.getCause()+" "+e.getMessage());
+            return false;
+        }
+        return true;
+    }
+    //修改老师电话，找到id对应的老师，按phone更新
+    public boolean updateTeacherPhoneById(int id,String phone){
+        //看老师是否存在，不存在报错
+        if(teacherDao.findById(id)==null){
+            System.out.println("修改错误，该老师不存在");
+            return false;
+        }
+        //通过id找到，修改对应信息
+        try {
+            TeacherEntity teacherEntity=teacherDao.findById(id);
+            teacherEntity.setPhone(phone);
+            teacherDao.save(teacherEntity);
+        }
+        catch (Exception e){
+            System.out.println("修改老师电话出错"+e.getCause()+" "+e.getMessage());
+            return false;
+        }
+        return true;
+    }
 }
